Re-organize to prepare to add back v2 support
