+++ lib/spack/spack/ci/common.py
@@ -36,6 +36,7 @@ from spack import traverse
 from spack.reporters import CDash, CDashConfiguration
 from spack.reporters.cdash import SPACK_CDASH_TIMEOUT
 from spack.reporters.cdash import build_stamp as cdash_build_stamp
+from spack.url_buildcache import create_url_buildcache_entry
 
 IS_WINDOWS = sys.platform == "win32"
 SPACK_RESERVED_TAGS = ["public", "protected", "notary"]
@@ -179,7 +180,9 @@ def write_pipeline_manifest(specs, src_prefix, dest_prefix, output_file):
     for release_spec in specs:
         release_spec_dag_hash = release_spec.dag_hash()
         # TODO: This assumes signed version of the spec
-        cache_entry = bindist.create_urlbuildcacheentry()
+        cache_entry = create_url_buildcache_entry(
+            layout_version=bindist.CURRENT_BUILD_CACHE_LAYOUT_VERSION
+        )
         buildcache_copies[release_spec_dag_hash] = {
             "src": cache_entry.compute_remote_spec_url(release_spec, src_prefix, signed=True),
             "dest": cache_entry.compute_remote_spec_url(release_spec, dest_prefix, signed=True),
