+++ lib/spack/spack/test/cmd/buildcache.py
@@ -26,6 +26,7 @@ import spack.util.url as url_util
 import spack.util.web as web_util
 from spack.installer import PackageInstaller
 from spack.paths import test_path
+from spack.url_buildcache import create_url_buildcache_entry
 
 buildcache = spack.main.SpackCommand("buildcache")
 install = spack.main.SpackCommand("install")
@@ -97,7 +98,9 @@ def tests_buildcache_create_env(
 
     mirror_url = f"file://{tmpdir.strpath}"
 
-    cache_entry = spack.binary_distribution.create_urlbuildcacheentry()
+    cache_entry = create_url_buildcache_entry(
+        layout_version=spack.binary_distribution.CURRENT_BUILD_CACHE_LAYOUT_VERSION
+    )
     cache_entry.initialize_from_spec_and_mirror(spec, mirror_url)
 
     exists = cache_entry.exists()
@@ -262,7 +265,9 @@ def test_buildcache_sync(
         shutil.rmtree(dest_mirror_dir)
 
         def manifest_insert(manifest, spec, dest_url):
-            cache_entry = spack.binary_distribution.create_urlbuildcacheentry()
+            cache_entry = create_url_buildcache_entry(
+                layout_version=spack.binary_distribution.CURRENT_BUILD_CACHE_LAYOUT_VERSION
+            )
             manifest[spec.dag_hash()] = {
                 "src": cache_entry.compute_remote_spec_url(spec, src_mirror_url, signed=False),
                 "dest": cache_entry.compute_remote_spec_url(spec, dest_url, signed=False),
@@ -316,7 +321,9 @@ def test_buildcache_create_install(
     mirror_url = f"file://{tmpdir.strpath}"
 
     spec = spack.concretize.concretize_one(pkg)
-    cache_entry = spack.binary_distribution.create_urlbuildcacheentry()
+    cache_entry = create_url_buildcache_entry(
+        layout_version=spack.binary_distribution.CURRENT_BUILD_CACHE_LAYOUT_VERSION
+    )
     cache_entry.initialize_from_spec_and_mirror(spec, mirror_url)
     spec_path = os.path.join(
         str(tmpdir), *cache_entry.get_relative_spec_components(spec, ".spec.json")
